agrichain/backend/requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
aiosqlite==0.20.0

# Caching / Session
redis==5.0.4

# WhatsApp / Twilio
twilio==9.0.5

# LangGraph + LangChain
langgraph==0.3.34
langchain==0.3.27
langchain-google-genai==2.1.12

# QR Code
qrcode[pil]==7.4.2
Pillow==10.3.0

# Validation
pydantic==2.12.5
pydantic-settings==2.11.0
email-validator==2.1.1
phonenumbers==8.13.37

# HTTP Client
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
apscheduler==3.10.4

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
